# HelpDeskUI/swagger.json
{
    "openapi": "3.0.4",
    "info": {
        "title": "HelpDeskAPI",
        "version": "v1"
    },
    "paths": {
...
        "/Ticket/assign": {
            "post": {
                "tags": [
                    "Ticket"
                ],
                "requestBody": {
                    "content": {
                        "application/json-patch+json": {
                            "schema": {
                                "$ref": "#/components/schemas/AssignTicketRequest"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AssignTicketRequest"
                            }
                        },
                        "text/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AssignTicketRequest"
                            }
                        },
                        "application/*+json": {
                            "schema": {
                                "$ref": "#/components/schemas/AssignTicketRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/BooleanResponse"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BooleanResponse"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BooleanResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/BooleanResponse"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BooleanResponse"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BooleanResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/Ticket/list": {
            "post": {
                "tags": [
                    "Ticket"
                ],
                "requestBody": {
                    "content": {
                        "application/json-patch+json": {
                            "schema": {
                                "$ref": "#/components/schemas/GetTicketsRequest"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GetTicketsRequest"
                            }
                        },
                        "text/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GetTicketsRequest"
                            }
                        },
                        "application/*+json": {
                            "schema": {
                                "$ref": "#/components/schemas/GetTicketsRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/ModelTicketModelWrapperResponse"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ModelTicketModelWrapperResponse"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ModelTicketModelWrapperResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/BooleanResponse"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BooleanResponse"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BooleanResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/Ticket/{ticketId}": {
            "delete": {
                "tags": [
                    "Ticket"
                ],
                "parameters": [
                    {
                        "name": "ticketId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/BooleanResponse"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BooleanResponse"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BooleanResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/BooleanResponse"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BooleanResponse"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BooleanResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/Ticket": {
            "post": {
                "tags": [
                    "Ticket"
                ],
                "requestBody": {
                    "content": {
                        "application/json-patch+json": {
                            "schema": {
                                "$ref": "#/components/schemas/InsertTicketRequest"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/InsertTicketRequest"
                            }
                        },
                        "text/json": {
                            "schema": {
                                "$ref": "#/components/schemas/InsertTicketRequest"
                            }
                        },
                        "application/*+json": {
                            "schema": {
                                "$ref": "#/components/schemas/InsertTicketRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/BooleanResponse"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BooleanResponse"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BooleanResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/BooleanResponse"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BooleanResponse"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BooleanResponse"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Ticket"
                ],
                "requestBody": {
                    "content": {
                        "application/json-patch+json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateTicketRequest"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateTicketRequest"
                            }
                        },
                        "text/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateTicketRequest"
                            }
                        },
                        "application/*+json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateTicketRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/BooleanResponse"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BooleanResponse"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BooleanResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/BooleanResponse"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BooleanResponse"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BooleanResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "AssignTicketRequest": {
                "required": [
                    "ticketId",
                    "userId"
                ],
                "type": "object",
                "properties": {
                    "ticketId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "userId": {
                        "type": "string",
                        "format": "uuid"
                    }
                },
                "additionalProperties": false
            },
            "BooleanResponse": {
                "type": "object",
                "properties": {
                    "errorCode": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "errorMessage": {
                        "type": "string",
                        "nullable": true
                    },
                    "result": {
                        "type": "boolean"
                    }
                },
                "additionalProperties": false
            },
            "Column": {
                "enum": [
                    "CreatedAt",
                    "FirstName",
                    "LastName",
                    "Status",
                    "UpdatedAt"
                ],
                "type": "string"
            },
...
            "GetTicketsRequest": {
                "required": [
                    "skip",
                    "sortColumns",
                    "take"
                ],
                "type": "object",
                "properties": {
                    "skip": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "sortColumns": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ModelSortColumn"
                        }
                    },
                    "status": {
                        "$ref": "#/components/schemas/TicketStatus"
                    },
                    "take": {
                        "type": "integer",
                        "format": "int32"
                    }
                },
                "additionalProperties": false
            },
            "InsertTicketRequest": {
                "required": [
                    "ticket"
                ],
                "type": "object",
                "properties": {
                    "ticket": {
                        "$ref": "#/components/schemas/ModelTicket"
                    }
                },
                "additionalProperties": false
            },
...
            "ModelSortColumn": {
                "required": [
                    "ascending",
                    "column"
                ],
                "type": "object",
                "properties": {
                    "ascending": {
                        "type": "boolean"
                    },
                    "column": {
                        "$ref": "#/components/schemas/Column"
                    }
                },
                "additionalProperties": false
            },
            "ModelTicket": {
                "required": [
                    "createdAt",
                    "description",
                    "id",
                    "status"
                ],
                "type": "object",
                "properties": {
                    "assignedUser": {
                        "$ref": "#/components/schemas/User"
                    },
                    "assignedUserId": {
                        "type": "string",
                        "format": "uuid",
                        "nullable": true
                    },
                    "createdAt": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "description": {
                        "minLength": 1,
                        "type": "string"
                    },
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "status": {
                        "$ref": "#/components/schemas/TicketStatus"
                    },
                    "updatedAt": {
                        "type": "integer",
                        "format": "int64",
                        "nullable": true
                    }
                },
                "additionalProperties": false
            },
            "ModelTicketModelWrapper": {
                "required": [
                    "results",
                    "total"
                ],
                "type": "object",
                "properties": {
                    "results": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ModelTicket"
                        }
                    },
                    "total": {
                        "type": "integer",
                        "format": "int32"
                    }
                },
                "additionalProperties": false
            },
            "ModelTicketModelWrapperResponse": {
                "type": "object",
                "properties": {
                    "errorCode": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "errorMessage": {
                        "type": "string",
                        "nullable": true
                    },
                    "result": {
                        "$ref": "#/components/schemas/ModelTicketModelWrapper"
                    }
                },
                "additionalProperties": false
            },
...
            "Ticket": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "assignedUserId": {
                        "type": "string",
                        "format": "uuid",
                        "nullable": true
                    },
                    "createdAt": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "nullable": true
                    },
                    "updatedAt": {
                        "type": "integer",
                        "format": "int64",
                        "nullable": true
                    },
                    "assignedUser": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "additionalProperties": false
            },
            "TicketStatus": {
                "enum": [
                    "Closed",
                    "InProgress",
                    "Open"
                ],
                "type": "string"
            },
            "UpdateTicketRequest": {
                "required": [
                    "description",
                    "id",
                    "status"
                ],
                "type": "object",
                "properties": {
                    "description": {
                        "minLength": 1,
                        "type": "string"
                    },
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "status": {
                        "$ref": "#/components/schemas/TicketStatus"
                    }
                },
                "additionalProperties": false
            },
            "User": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "firstName": {
                        "type": "string",
                        "nullable": true
                    },
                    "lastName": {
                        "type": "string",
                        "nullable": true
                    },
                    "password": {
                        "type": "string",
                        "nullable": true
                    },
                    "username": {
                        "type": "string",
                        "nullable": true
                    },
                    "ticket": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Ticket"
                        },
                        "nullable": true
                    }
                },
                "additionalProperties": false
            }
        }
    }
}